requirements.txt
django
pycryptodome
